+++ CHANGELOG.md
@@ -44,6 +44,8 @@
     * CLI better source_type determining
     * web module boosts and fixes
     * fix: when reprocessing (ex: invalid lines) after loading, file will get cleared no more
+    * fix: CIDR when WHOIS returns directly CIDR instead of prefix. Currently, we do not guarantee prefix will be in the form "... - ...", it may has directly CIDR form if WHOIS tells us so
+    * fix 1.3.6: file name duplicate when reprocessing
 
 ## 1.3.1 (2020-01-31)
 * thread processing